+++ polimorfo/datasets/instancecoco.py
@@ -17,9 +17,14 @@ class InstanceCoco(CocoDataset):
 
     """
 
-    def add_annotations(self, img_id: int, labels: np.ndarray,
-                        boxes: np.ndarray, masks: np.ndarray,
-                        scores: np.ndarray) -> List[int]:
+    def add_annotations(
+        self,
+        img_id: int,
+        labels: np.ndarray,
+        boxes: np.ndarray,
+        scores: np.ndarray,
+        masks: np.ndarray = None,
+    ) -> List[int]:
         """
         add the annotations from the given results
 
@@ -27,13 +32,13 @@ class InstanceCoco(CocoDataset):
             img_id (int): the idx of the image to add the annotations
             labels (np.ndarray): an array with the label predicted
             boxes (np.ndarray): an array of shape (n_labels, 4)
-            masks (np.ndarray): an array of shape (Height, Width, n_labels)
             scores (np.ndarray): an array with the scores of the predictions
+            masks (np.ndarray): an array of shape (Height, Width, n_labels) (Optional: None is your saving only boxes)
 
         Raises:
             ValueError: labels.shape should equal to 1, (n_labels,)
             ValueError: boxes.shape should equal to 2 (n_labels, 4)
-            ValueError: masks.shape should equal to 3, (Height, Width, n_labels)
+            ValueError: masks.shape should equal to 3, (n_labels, Height, Width, n_labels)
             ValueError: scores.shape should equal to 1, (n_labels,)
 
         Returns:
@@ -46,11 +51,12 @@ class InstanceCoco(CocoDataset):
         if len(boxes.shape) != 2:
             raise ValueError('boxes.shape should equal to 2 (n_labels, 4)')
 
-        masks = masks.squeeze()
+        if masks is not None:
+            masks = masks.squeeze()
 
-        if len(masks.shape) != 3:
-            raise ValueError(
-                'masks.shape should equal to 3, (Height, Width, n_labels)')
+            if len(masks.shape) != 3:
+                raise ValueError(
+                    'masks.shape should equal to 3, (n_labels, Height, Width)')
 
         if len(scores.shape) != 1:
             raise ValueError('scores.shape should equal to 1, (n_labels,)')
@@ -63,13 +69,20 @@ class InstanceCoco(CocoDataset):
             # convert box to
             x0, y0, x1, y1 = boxes[i]
             w, h = x1 - x0, y1 - y0
-            bbox = [x0, y0, w, h]
+            area = int(w * h)
+            if area == 0:
+                continue
+
+            bbox = [float(x0), float(y0), float(w), float(h)]
             # create the polygons
-            mask = masks[..., i]
-            polygons = maskutils.mask_to_polygon(mask)
-            area = maskutils.area(mask)
-            score = scores[i]
+            if masks is not None:
+                mask = masks[i, ...]
+                polygons = maskutils.mask_to_polygon(mask)
+            else:
+                polygons = []
+
+            score = float(scores[i])
             annotation_ids.append(
-                self.add_annotation(img_id, cat_id, polygons, area, bbox, 0,
-                                    score))
+                self.add_annotation(int(img_id), int(cat_id), polygons, area,
+                                    bbox, 0, score))
         return annotation_ids