+++ polimorfo/utils/maskutils.py
@@ -19,10 +19,27 @@ def mask_to_polygon(mask, min_score=0.5):
                                 offset=(-1, -1))
     polygons = polygons[0] if len(polygons) == 2 else polygons[1]
     polygons = [polygon.flatten().tolist() for polygon in polygons]
+    # add filter to remove invalid polygons
+    polygons = [polygon for polygon in polygons if len(polygon) >= 8]
     return polygons
 
 
 def polygons_to_mask(polygons, height, width):
+    """convert polygons to mask. Filter all the polygons with less than 4 points
+
+    Args:
+        polygons ([type]): [description]
+        height ([type]): [description]
+        width ([type]): [description]
+
+    Returns:
+        [type]: [description]
+    """
+    polygons = [polygon for polygon in polygons if len(polygon) >= 8]
+
+    if len(polygons) == 0:
+        return np.zeros((height, width), np.uint8)
+
     rle = mask_util.frPyObjects(polygons, height, width)
     rle = mask_util.merge(rle)
     return mask_util.decode(rle)[:, :]
@@ -42,7 +59,7 @@ def bbox(polygons, height, width):
     return bbox
 
 
-def coco_poygons_to_mask(segmentations, height, width):
+def coco_poygons_to_mask(segmentations, height, width) -> np.ndarray:
     masks = []
     for polygons in segmentations:
         mask = polygons_to_mask(polygons, height, width)