+++ tests/datasets/test_coco.py
@@ -2,7 +2,11 @@ import pytest
 from pytest import fixture
 from pathlib import Path
 import shutil
-from polimorfo.datasets import CocoDataset, ExportFormat
+from polimorfo.datasets import CocoDataset
+import os
+import numpy as np
+from PIL import Image
+from tqdm import tqdm
 
 BASE_PATH = Path(__file__).parent.parent / 'data'
 
@@ -26,12 +30,14 @@ def test_load_coco(dataset_file):
 def test_categories_images_count(coco_test):
     images_count = coco_test.cats_images_count()
     assert len(images_count) == 3
+    print(images_count)
     assert images_count == {'toaster': 217, 'hair drier': 189, 'bear': 960}
 
 
 def test_categories_annotations_count(coco_test):
     images_count = coco_test.cats_annotations_count()
     assert len(images_count) == 3
+    print(images_count)
     assert images_count == {'toaster': 225, 'hair drier': 198, 'bear': 1294}
 
 
@@ -50,13 +56,27 @@ def test_dumps(coco_test):
 
 def test_dump_segmentation(coco_test):
     out_path = BASE_PATH / 'segments'
-    coco_test.dump(out_path, ExportFormat.segmentation)
+    coco_test.dump(out_path, CocoDataset.ExportFormat.segmentation)
     assert len(list(out_path.glob('*.png'))) > 0
     shutil.rmtree(out_path.as_posix())
 
 
+def test_save_segmentation_masks(coco_test):
+    out_path = BASE_PATH / 'segments'
+    coco_test.save_segmentation_masks(out_path, [23], {23: 24})
+    assert len(list(out_path.glob('*.png'))) > 0
+    distinct_values = set()
+    for png_path in tqdm(list(out_path.glob('*.png'))):
+        img = np.array(Image.open(png_path))
+        distinct_values = distinct_values.union(set(np.unique(img)))
+
+    assert distinct_values == {0, 24}
+    shutil.rmtree(out_path.as_posix())
+
+
 def test_create_dataset():
-    ds = CocoDataset()
+    annotations_path = BASE_PATH / 'new_coco.json'
+    ds = CocoDataset(annotations_path)
 
     cat_id = ds.add_category('dog', 'animal')
     img_id = ds.add_image((BASE_PATH / 'test_nodamage.jpg').as_posix())
@@ -65,9 +85,17 @@ def test_create_dataset():
     assert len(ds) == 1
     assert len(ds.anns) == 1
 
+    ds.dump()
+    assert annotations_path.exists()
+
+    with open(annotations_path, 'r') as f:
+        assert len(f.readline()) > 0
+
+    os.remove(annotations_path)
+
 
 def test_create_dataset_existing():
-    ds = CocoDataset()
+    ds = CocoDataset(BASE_PATH / 'new_coco.json')
 
     cat_id = ds.add_category('dog', 'animal')
     img_id = ds.add_image((BASE_PATH / 'test_nodamage.jpg').as_posix())
@@ -85,7 +113,7 @@ def test_create_dataset_existing():
 
 
 def test_remove_categories():
-    ds = CocoDataset()
+    ds = CocoDataset(BASE_PATH / 'new_coco.json')
     cat_id = ds.add_category('dog', 'animal')
     assert len(ds.cats) == 1
     ds.remove_categories([cat_id])
@@ -94,7 +122,7 @@ def test_remove_categories():
 
 
 def test_remove_categories_and_annotations():
-    ds = CocoDataset()
+    ds = CocoDataset(BASE_PATH / 'new_coco.json')
     cat_id = ds.add_category('dog', 'animal')
     img_id = ds.add_image((BASE_PATH / 'test_nodamage.jpg').as_posix())
     ds.add_annotation(img_id, cat_id, [1, 2, 3, 4, 5], 10, [0, 0, 256, 256], 0)
@@ -129,4 +157,4 @@ def test_copy(coco_test: CocoDataset):
 def test_update_images_path(coco_test: CocoDataset):
     coco_test.update_images_path(lambda x: Path(x).name)
     coco_test.reindex()
-    assert coco_test.imgs[1]['file_name'] == '000000410627.jpg'
+    assert coco_test.imgs[1]['file_name'] == '000000001442.jpg'